+++ frontend/src/components/TimeEntry/AddEntryForm.js
@@ -20,7 +20,7 @@ import {Button, Col, Form, Input, message, Radio, Row, Select, Space} from 'antd
 import PropTypes from 'prop-types';
 import TitleSection from '../Title/TitleSection';
 import './AddEntryForm.less';
-var _ = require('lodash');
+import {hoursOptions, radioDisabledClassName} from '../../utils/utils';
 
 const {Option} = Select;
 const {TextArea} = Input;
@@ -118,19 +118,6 @@ const AddEntryForm = ({date, form, timeSheets, onSuccess, onCancel, numberOfHour
     }
   };
 
-  const hoursOptions = (numberOfHoursForDay) => {
-    const hourDisabled = (hour) => {
-      return parseInt(hour) + parseInt(numberOfHoursForDay) > 8;
-    };
-    return _.range(1, 9, 1).map(i => <Option key={`option-hour-${i}`} disabled={hourDisabled(i)} value={i} >{i}</Option>);
-  };
-
-  const disabledClassName = (disabled) => {
-    if(disabled) {
-      return 'tk-radio-disabled';
-    }
-  };
-
   return (
     <div className="tk_ModalBottom">
       <Form
@@ -168,8 +155,8 @@ const AddEntryForm = ({date, form, timeSheets, onSuccess, onCancel, numberOfHour
                 return (
                   <Form.Item name="timeUnit" label="Logged time:" rules={[{required: true}]}>
                     <Radio.Group>
-                      <Radio value="DAY" className={disabledClassName(dayDisabled)} disabled={dayDisabled}>Day</Radio>
-                      <Radio value="HALFDAY" className={disabledClassName(halfDayDisabled)} disabled={halfDayDisabled}>Half-day</Radio>
+                      <Radio value="DAY" className={radioDisabledClassName(dayDisabled)} disabled={dayDisabled}>Day</Radio>
+                      <Radio value="HALFDAY" className={radioDisabledClassName(halfDayDisabled)} disabled={halfDayDisabled}>Half-day</Radio>
                       <Radio value="HOURLY" disabled={hourDisabled}>Hours</Radio>
                     </Radio.Group>
                   </Form.Item>
