+++ frontend/src/components/TimeEntry/EditEntryForm.js
@@ -21,7 +21,7 @@ import PropTypes from 'prop-types';
 import TitleSection from '../Title/TitleSection';
 import moment from 'moment';
 import {computeNumberOfHours} from '../../utils/momentUtils';
-var _ = require('lodash');
+import {hoursOptions, radioDisabledClassName} from '../../utils/utils';
 
 const {Option} = Select;
 const {TextArea} = Input;
@@ -141,19 +141,6 @@ const EditEntryForm = ({date, form, timeSheets, onSuccess, onCancel, entry, numb
     return timeSheets.find(timeSheet => !!timeSheet.entries.find(entry => entry.id === entryId)).id;
   };
 
-  const hoursOptions = (numberOfHoursForDay, entryDuration) => {
-    const hourDisabled = (hour) => {
-      return parseInt(hour) + (parseInt(numberOfHoursForDay) - entryDuration) > 8;
-    };
-    return _.range(1, 9, 1).map(i => <Option key={`option-hour-${i}`} disabled={hourDisabled(i)} value={i} >{i}</Option>);
-  };
-
-  const disabledClassName = (disabled) => {
-    if(disabled) {
-      return 'tk-radio-disabled';
-    }
-  };
-
   return (
     <div className="tk_ModalBottom">
       <Form
@@ -192,8 +179,8 @@ const EditEntryForm = ({date, form, timeSheets, onSuccess, onCancel, entry, numb
                 return (
                   <Form.Item name="timeUnit" label="Logged time:" rules={[{required: true}]}>
                     <Radio.Group>
-                      <Radio value="DAY" className={disabledClassName(dayDisabled)} disabled={dayDisabled}>Day</Radio>
-                      <Radio value="HALFDAY" className={disabledClassName(halfDayDisabled)} disabled={halfDayDisabled}>Half-day</Radio>
+                      <Radio value="DAY" className={radioDisabledClassName(dayDisabled)} disabled={dayDisabled}>Day</Radio>
+                      <Radio value="HALFDAY" className={radioDisabledClassName(halfDayDisabled)} disabled={halfDayDisabled}>Half-day</Radio>
                       <Radio value="HOURLY" disabled={hourDisabled}>Hours</Radio>
                     </Radio.Group>
                   </Form.Item>
